append_claim, append_evidenceに対応したテストコードの修正
