+++ src/__tests__/services/point.services.test.ts
@@ -4,7 +4,7 @@ import {point_slice} from 'stores/slices/point';
 import {evidence_slice} from 'stores/slices/evidence';
 import {is_mPoint, mPoint,mPointSignature} from 'models/mPoint';
 import { mEvidenceSignature } from 'models/mEvidence';
-import {generate_point,point_add_child,append_claim,append_point_to_part, append_sibling_point, append_point_child} from 'services/point';
+import {generate_point,point_add_child,append_point_to_part, append_sibling_point, append_point_child} from 'services/point';
 import {generate_part, part_add_child} from 'services/part';
 import { mClaim, mClaimSignature } from 'models/mClaim';
 import { generate_claim } from 'services/claim';
@@ -95,60 +95,6 @@ test('point_add_child: Point',()=>{
   expect(store.getState().point.entities[generated.id]).toMatchObject(expected_result);
   expect(store.getState().point.entities[modified.id]).toBeTruthy();
   const parent_in_redux=get_from_id(generated.id);
-  expect(is_mPoint(parent_in_redux)).toBeTruthy();
-  if(!is_mPoint(parent_in_redux)) return;
-  expect(parent_in_redux.contents).toContain(modified.id);
-});
-
-test('append_claim: Point',()=>{
-  const part_id=generate_parents()[2];
-  const parent=generate_point(part_id);
-  const modified=append_claim(parent.id);
-  expect(modified).not.toBeUndefined();
-  if(modified===undefined) return;
-  const expected_result:Omit<mClaim,'id'> = {
-    type_signature: mClaimSignature,
-    parent: parent.id,
-  };
-  expect(store.getState().claim.entities[modified.id]).toMatchObject(expected_result);
-  const parent_in_redux=get_from_id(parent.id);
-  expect(is_mPoint(parent_in_redux)).toBeTruthy();
-  if(!is_mPoint(parent_in_redux)) return;
-  expect(parent_in_redux.contents).toContain(modified.id);
-});
-
-test('append_claim: Part',()=>{
-  const part_id=generate_parents()[2];
-  const modified=append_claim(part_id);
-  expect(modified).not.toBeUndefined();
-  if(modified===undefined) return;
-  const result_in_redux=store.getState().claim.entities[modified.id];
-  expect(result_in_redux).toBeTruthy();
-  if(result_in_redux===undefined) return;
-  const part_in_redux=get_from_id(part_id);
-  expect(is_mPart(part_in_redux)).toBe(true);
-  if(!is_mPart(part_in_redux)) return;
-  expect(part_in_redux.contents).toBeTruthy();
-  if(!part_in_redux.contents) return;
-  expect(result_in_redux.parent).toBe(part_in_redux.contents[part_in_redux.contents.length-1]);
-  const parent_in_redux=get_from_id(result_in_redux.parent);
-  expect(is_mPoint(parent_in_redux)).toBeTruthy();
-  if(!is_mPoint(parent_in_redux)) return;
-  expect(parent_in_redux.contents).toContain(modified.id);
-});
-
-test('append_claim: Claim',()=>{
-  const part_id=generate_parents()[2];
-  const parent=generate_point(part_id);
-  const claim=generate_claim(parent.id);
-  const modified=append_claim(claim.id);
-  expect(modified).not.toBeUndefined();
-  if(modified===undefined) return;
-  const result_in_redux=store.getState().claim.entities[modified.id];
-  expect(result_in_redux).toBeTruthy();
-  if(result_in_redux===undefined) return;
-  expect(result_in_redux.parent).toBe(parent.id);
-  const parent_in_redux=get_from_id(parent.id);
   expect(is_mPoint(parent_in_redux)).toBeTruthy();
   if(!is_mPoint(parent_in_redux)) return;
   expect(parent_in_redux.contents).toContain(modified.id);
